ENH: use Pythran to speedup somersd and _tau_b (#14308)
